refactor(db): consolidate database.py into extensions.py and update imports
