+++ backend/app/extensions.py
@@ -7,3 +7,4 @@ from flask_jwt_extended import JWTManager
 db = SQLAlchemy()
 migrate = Migrate()
 jwt = JWTManager()
+Base = db.Model
