+++ backend/app/models/models.py
@@ -2,7 +2,7 @@
 
 from sqlalchemy import Enum, CheckConstraint
 from werkzeug.security import generate_password_hash, check_password_hash
-from backend.database import db
+from app.extensions import db
 
 # Allowed event statuses
 EVENT_STATUSES = ("drafting", "published", "cancelled", "completed")
